backend/pyproject.toml
[project]
name = "smart-mirror-backend"
version = "0.1.0"
description = "Smart mirror backend"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
    "flask>=3.0",
    "flask-cors>=4.0",
    "face-recognition>=1.3.0",
    "dlib>=19.24,<20",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    # face_recognition_models imports pkg_resources; setuptools 82+ drops pkg_resources
    "setuptools>=69,<81",
    "supabase>=2.0,<2.10",
    "google-genai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["app.py", "db.py", "auth.py", "face_store.py"]
